Finished memory #4 and fixed a bug in memories.gd
If the player was touching the mushroom, and pressed the down arrow, a message was printed in the log. Theoritically, the player could crash the game by spamming the down arrow.
Now, it only prints a message if the are actually in a memory cutscene, and if they are not in level 4 (since it autoplays and they can, once again, spam down key with no consequences).