Fixed a new bug with memories.gd, deleted the "More coming soon" frame from memory 4, and added the cloudy effect to the memories.
I found out that I accidently added a "not" infront of the cutscene_in_progress check, causing issues with the non-autoplay memories. Fixed now!

I also realized that, in my relief to be done, I forgot to delete the "more coming soon" and add the cloudy effect to the memories. That's down too, now.